requirements.txt
uvicorn
fastapi
motor
python-dotenv